e2e dataset details
